refactor handleSubmitForm
